+++ libs/idp-server-core-extension-ciba/src/main/java/org/idp/server/core/extension/ciba/CibaRequestContext.java
@@ -20,10 +20,10 @@ import java.util.HashMap;
 import java.util.List;
 import org.idp.server.core.extension.ciba.request.BackchannelAuthenticationRequest;
 import org.idp.server.core.extension.ciba.request.BackchannelAuthenticationRequestIdentifier;
-import org.idp.server.core.extension.ciba.user.UserHint;
-import org.idp.server.core.extension.ciba.user.UserHintRelatedParams;
-import org.idp.server.core.extension.ciba.user.UserHintType;
 import org.idp.server.core.openid.authentication.AuthenticationInteractionType;
+import org.idp.server.core.openid.identity.hint.UserHint;
+import org.idp.server.core.openid.identity.hint.UserHintRelatedParams;
+import org.idp.server.core.openid.identity.hint.UserHintType;
 import org.idp.server.core.openid.oauth.clientauthenticator.BackchannelRequestContext;
 import org.idp.server.core.openid.oauth.clientauthenticator.BackchannelRequestParameters;
 import org.idp.server.core.openid.oauth.configuration.AuthorizationServerConfiguration;
