+++ libs/idp-server-core/src/main/java/org/idp/server/core/openid/identity/hint/UserHintRelatedParams.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.idp.server.core.extension.ciba.user;
+package org.idp.server.core.openid.identity.hint;
 
 import java.util.Map;
 import java.util.function.BiConsumer;
