+++ libs/idp-server-core/src/main/java/org/idp/server/core/openid/identity/hint/UserHintResolver.java
@@ -14,15 +14,17 @@
  * limitations under the License.
  */
 
-package org.idp.server.core.extension.ciba.user;
+package org.idp.server.core.openid.identity.hint;
 
 import org.idp.server.core.openid.identity.User;
-import org.idp.server.platform.type.Pairs;
+import org.idp.server.core.openid.identity.repository.UserQueryRepository;
+import org.idp.server.platform.multi_tenancy.tenant.Tenant;
 
-public interface LoginHintMatcher {
-  boolean matches(String hint);
+public interface UserHintResolver {
 
-  Pairs<String, String> extractHints(String hint);
-
-  User resolve(Pairs<String, String> hints);
+  User resolve(
+      Tenant tenant,
+      UserHint userHint,
+      UserHintRelatedParams userHintRelatedParams,
+      UserQueryRepository userQueryRepository);
 }
