+++ libs/idp-server-use-cases/src/main/java/org/idp/server/usecases/application/enduser/OAuthFlowEntryService.java
@@ -20,9 +20,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import org.idp.server.core.extension.ciba.user.LoginHintResolver;
-import org.idp.server.core.extension.ciba.user.UserHint;
-import org.idp.server.core.extension.ciba.user.UserHintRelatedParams;
 import org.idp.server.core.openid.authentication.AuthSessionId;
 import org.idp.server.core.openid.authentication.AuthSessionValidator;
 import org.idp.server.core.openid.authentication.Authentication;
@@ -49,6 +46,9 @@ import org.idp.server.core.openid.identity.UserRegistrator;
 import org.idp.server.core.openid.identity.event.UserLifecycleEvent;
 import org.idp.server.core.openid.identity.event.UserLifecycleEventPublisher;
 import org.idp.server.core.openid.identity.event.UserLifecycleType;
+import org.idp.server.core.openid.identity.hint.LoginHintResolver;
+import org.idp.server.core.openid.identity.hint.UserHint;
+import org.idp.server.core.openid.identity.hint.UserHintRelatedParams;
 import org.idp.server.core.openid.identity.repository.UserCommandRepository;
 import org.idp.server.core.openid.identity.repository.UserQueryRepository;
 import org.idp.server.core.openid.oauth.*;
